requirements.txt
aiohttp==3.7.2
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
chardet==3.0.4
classcharts.py==0.1.1
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.1
MarkupSafe==1.1.1
multidict==5.0.0
PyMySQL==0.10.1
python-decouple==3.3
soupsieve==2.0.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
yarl==1.6.2
